This is a file based on the problems STRESS @HTML
Do not over work yourself or you may be able to fall in love with the wrong piece of work that is in the work environement by becoming strict it will lead to producing a website that is in tune to this. The code process the CSS and HTML plus some of the latest use of Notepad with Visual Studio Code you are drawn to not be at a struggle but with a Laptop running at this state it is good for the coding in the Training. @Quackit @W3 is a strong form of why you arent too never try being in a away from the Mathematic choice of  this. So MatLab or not Creating a device for the future you based on access countless amount of links and disfiguring the page to look the way you  want it. HTML is a powerful webpage but that is not alll as there are other that can lead people to not use it how can you not be a user but search for something that will promote this in the fute for those that nneds and will probably need. This is a pyrmaid of high HTML and CSS. I will study more about the problems that are there for some of the educated people in the program. A program that is being develop will be work on this.