Feeling is Another website. @HTML @CSS #Notepad
How can I be able to use JavaScript when there is a multiple amount of languages that are in the Linux program so one ought to change the path of programing and make this a unique experience so far I can  only do so much that this is all about create to much contain each day that is not for me but for others to explore and make sure that this is better for them in the next day. C++ is not install and work be install for a while so how can I  differ from others. That may use this please let me know because  it is tiring. #VISUAL STUDIO CODE